ICU-7467 Backwards-compatible deserialization support and assorted minor fixes.  All unit tests in core-tests are passing!
X-SVN-Rev: 39847

+++ icu4j/main/classes/core/src/com/ibm/icu/impl/number/FormatQuantity1.java
@@ -277,8 +277,8 @@ public class FormatQuantity1 implements FormatQuantity {
     long fingerprint = 0;
     fingerprint ^= lOptPos;
     fingerprint ^= (lReqPos << 16);
-    fingerprint ^= (rReqPos << 32);
-    fingerprint ^= (rOptPos << 48);
+    fingerprint ^= ((long) rReqPos << 32);
+    fingerprint ^= ((long) rOptPos << 48);
     return fingerprint;
   }
